POS Mobile Application

+++ src/main/java/com/dinsaren/bbuappserver/repository/CategoryRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface CategoryRepository extends CrudRepository<Category, Long> {
+public interface CategoryRepository extends CrudRepository<Category, Integer> {
     List<Category> findAllByStatus(String status);
 }
