+++ src/main/java/com/dinsaren/bbuappserver/repository/ProductRepository.java
@@ -5,7 +5,8 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface ProductRepository extends CrudRepository<Product, Long> {
+public interface ProductRepository extends CrudRepository<Product, Integer> {
     List<Product> findAllByStatus(String status);
+
     List<Product> findAllByCategoryIdAndStatus(Integer categoryId, String status);
 }
